+++ Sources/PaymentButtons/PaymentButtonLabel.swift
@@ -71,28 +71,6 @@ public enum PaymentButtonLabel: String {
     }
 
     var position: Position {
-        switch self {
-        case .addMoneyWith,
-            .bookWith,
-            .buyWith,
-            .buyNowWith,
-            .checkoutWith,
-            .continueWith,
-            .contributeWith,
-            .orderWith,
-            .payLaterWith,
-            .payWith,
-            .reloadWith,
-            .rentWith,
-            .reserveWith,
-            .subscribeWith,
-            .supportWith,
-            .tipWith,
-            .topUpWith:
-            return .prefix
-
-        case .payLater:
-            return .suffix
-        }
+        return self == .payLater ? .suffix : .prefix
     }
 }