+++ CHANGELOG.md
@@ -3,10 +3,13 @@
 
 ## unreleased
 * PaymentButtons
-  * Add new size `.miniWithWordmark`
-  * Add new `PayPalButton` labels: `.addMoneyWith`, `.bookWith`, `.buyWith`, `.buyNowWith`, `.checkoutWith`, `.continueWith`, `.contributeWith`, `.orderWith`, `.payLaterWith`, `.payWith`, `.reloadWith`, `.rentWith`, `.reserveWith`, `.subscribeWith`, `.supportWith`, `.tipWith`, `.topUpWith`
+  * Add new `PaymentButtonSize` case `.miniWithWordmark`
+  * Add new `PayPalButton.Labels` cases: `.addMoneyWith`, `.bookWith`, `.buyWith`, `.buyNowWith`,
+  `.checkoutWith`, `.continueWith`, `.contributeWith`, `.orderWith`, `.payLaterWith`, `.payWith`,
+  `.reloadWith`, `.rentWith`, `.reserveWith`, `.subscribeWith`, `.supportWith`, `.tipWith`,
+  `.topUpWith`
 * Breaking Changes
-  * Remove `.checkout`, `.buyNow`, `.payLater`
+  * Remove `PayPalButton.Labels` cases: `.checkout`, `.buyNow`, `.payLater`
 
 ## 1.1.0 (2023-11-16)
 * PayPalNativePayments